Add tests for utility functions
